Crate miden_core

Modules

advice

chiplets

crypto

events

field

mast

MAST forest: a collection of procedures represented as Merkle trees.

operations

precompile

Precompile framework for deferred verification in the Miden VM.

prettier

program

proof

serde

utils

Macros

assert_matches

This is an implementation of std::assert_matches::assert_matches so it can be removed when that feature stabilizes upstream

Structs

Felt

A Felt backed by Plonky3’s Goldilocks field element.

LexicographicWord

A Word wrapper with lexicographic ordering.

Word

A unit of data consisting of 4 field elements.

Constants

EMPTY_WORD

Array of field elements representing word of ZEROs in the Miden base field.

FMP_ADDR

The address where the frame pointer is stored in memory.

FMP_INIT_VALUE

The initial value for the frame pointer, corresponding to the start address for procedure locals.

ONE

Field element representing ONE in the Miden base filed.

WORD_SIZE

Number of field elements in a word.

ZERO

Field element representing ZERO in the Miden base filed.