miden_core::field

Trait TwoAdicField

pub trait TwoAdicField: Field {
    const TWO_ADICITY: usize;

    // Required method
    fn two_adic_generator(bits: usize) -> Self;
}

A field which supplies information like the two-adicity of its multiplicative group, and methods for obtaining two-adic generators.

Required Associated Constants

const TWO_ADICITY: usize

The number of factors of two in this field’s multiplicative group.

Required Methods

fn two_adic_generator(bits: usize) -> Self

Returns a generator of the multiplicative group of order 2^bits. Assumes bits <= TWO_ADICITY, otherwise the result is undefined.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.