miden_core

Module operations

Modules

opcodes

Opcode patterns have the following meanings:

Structs

AssemblyOp

Contains information corresponding to an assembly instruction (only applicable in debug mode).

DebugVarInfo

Debug information for tracking a source-level variable.

Enums

DebugOptions

Options of the Debug decorator.

DebugVarLocation

Describes where a variable’s value can be found during execution.

Decorator

A set of decorators which can be executed by the VM.

Operation

The set of native VM basic block operations executable which take exactly one cycle to execute.

Type Aliases

DecoratorList

Vector consisting of a tuple of operation index (within a span block) and decorator at that index.