miden_core

Module program

Structs

Kernel

A list of exported kernel procedure hashes defining a VM kernel.

Program

An executable program for Miden VM.

ProgramInfo

A program information set consisting of its MAST root and set of kernel procedure roots used for its compilation.

StackInputs

Defines the initial state of the VM’s operand stack.

StackOutputs

Defines the final state of the VM’s operand stack at the end of program execution.

Enums

InputError

KernelError

OutputError

Constants

MIN_STACK_DEPTH

Represents: